requirements.txt
fastapi
uvicorn
prometheus-client
boto3
scikit-learn
prophet
numpy
pandas
python-dotenv
